move gcounter tests to tests/
